Tidy implemented properties
