+++ crates/rmf_site_editor/src/site/mod.rs
@@ -57,6 +57,9 @@ pub use group::*;
 pub mod headless_export;
 pub use headless_export::*;
 
+pub mod inclusion;
+pub use inclusion::*;
+
 pub mod lane;
 pub use lane::*;
