+++ crates/rmf_site_editor/src/site/modifier.rs
@@ -16,19 +16,14 @@
 */
 
 use crate::{
-    site::{
-        Affiliation, Element, Inclusion, IssueKey, NameInSite, Pending, Property, ScenarioModifiers,
-    },
+    site::{Affiliation, IssueKey, NameInSite, Property, ScenarioModifiers},
     Issue, ValidateWorkspace,
 };
 use bevy::{
-    ecs::{
-        component::Mutable,
-        system::{SystemParam, SystemState},
-    },
+    ecs::{component::Mutable, system::SystemParam},
     prelude::*,
 };
-use std::{collections::HashSet, fmt::Debug};
+use std::fmt::Debug;
 use uuid::Uuid;
 
 #[derive(Component, Debug, Default, Clone, Deref, DerefMut)]
@@ -37,75 +32,6 @@ pub struct Modifier<T: Property>(T);
 impl<T: Property> Modifier<T> {
     pub fn new(value: T) -> Self {
         Self(value)
-    }
-}
-
-impl Property for Inclusion {
-    fn get_fallback(_for_element: Entity, _in_scenario: Entity, _world: &mut World) -> Inclusion {
-        Inclusion::default()
-    }
-
-    fn on_new_element(
-        for_element: Entity,
-        in_scenario: Entity,
-        _value: Inclusion,
-        world: &mut World,
-    ) {
-        let mut scenario_state: SystemState<
-            Query<(Entity, &ScenarioModifiers<Entity>, &Affiliation<Entity>)>,
-        > = SystemState::new(world);
-        let scenarios = scenario_state.get_mut(world);
-
-        // Insert inclusion modifier into all root scenarios outside of the current tree as hidden
-        let mut current_root_entity: Entity = in_scenario;
-        while let Ok((_, _, parent_scenario)) = scenarios.get(current_root_entity) {
-            if let Some(parent_scenario_entity) = parent_scenario.0 {
-                current_root_entity = parent_scenario_entity;
-            } else {
-                break;
-            }
-        }
-        let mut root_scenarios = HashSet::<Entity>::new();
-        for (scenario_entity, _, parent_scenario) in scenarios.iter() {
-            if parent_scenario.0.is_some() || scenario_entity == current_root_entity {
-                continue;
-            }
-            root_scenarios.insert(scenario_entity);
-        }
-        for root in root_scenarios.iter() {
-            world.trigger(UpdateModifier::modify(
-                *root,
-                for_element,
-                Inclusion::Hidden,
-            ));
-        }
-    }
-
-    fn on_new_scenario<E: Element>(in_scenario: Entity, world: &mut World) {
-        let mut state: SystemState<(
-            Query<&Children>,
-            Query<(&Modifier<Inclusion>, &Affiliation<Entity>)>,
-            Query<Entity, (With<E>, Without<Pending>)>,
-        )> = SystemState::new(world);
-        let (children, task_modifiers, task_entities) = state.get_mut(world);
-
-        let have_task = Self::elements_with_modifiers(in_scenario, &children, &task_modifiers);
-
-        let mut target_tasks = HashSet::new();
-        for task_entity in task_entities.iter() {
-            if !have_task.contains(&task_entity) {
-                target_tasks.insert(task_entity);
-            }
-        }
-
-        for target in target_tasks.iter() {
-            // Mark all task modifiers as Hidden
-            world.trigger(UpdateModifier::modify(
-                in_scenario,
-                *target,
-                Inclusion::Hidden,
-            ));
-        }
     }
 }
